% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_template_images.R
\name{ct_malf_images}
\alias{ct_malf_images}
\alias{ct_malf_template_images}
\alias{ct_malf_template_masks}
\title{Get Image Filenames for templates}
\usage{
ct_malf_images()

ct_malf_template_images(...)

ct_malf_template_masks(...)
}
\arguments{
\item{...}{arguments to pass \code{\link{ct_malf_images}}}
}
\value{
List of filenames
}
\description{
Return a list of filenames for the images
}
